Change api path and better error handling
