+++ app/routes/challenges_routes.py
@@ -7,16 +7,15 @@
 
 '''
 from flask import Blueprint, request, jsonify
-from app.services import redis_client, leaderboard_manager, problem_manager, firestore_wrapper
+from app.services import redis_client, leaderboard_manager, problem_manager
 from constants import REDIS_ATTEMPTED_KEY, REDIS_SOLVED_KEY
-from datetime import datetime
 
 bp = Blueprint('challenges', __name__, url_prefix='/challenges')
 
 
 # Used by Leetcode bot to get a random problem of the week
-@bp.route('/problem/weekly', methods=['GET'])
-def get_weekly_problem():
+@bp.route('/problem', methods=['GET'])
+def get_problem():
     difficulty = request.args.get('difficulty')
     if not difficulty:
         return jsonify({'error': 'Difficulty not provided'}), 400
@@ -29,33 +28,6 @@ def get_weekly_problem():
         return jsonify({'error': 'No problem found'}), 404
     
     return jsonify(problem), 200
-
-# Used by Leetcode bot to get a random problem as per user request
-@bp.route('/problem/<difficulty>/request', methods=['GET'])
-def get_problem_user_request(difficulty):
-    if not difficulty:
-        return jsonify({'error': 'Difficulty not provided'}), 400
-    
-    if difficulty not in ['easy', 'medium', 'hard']:
-        return jsonify({'error': 'Invalid difficulty'}), 400
-
-    # select_problem_at_random() = no cache
-    problem = problem_manager.select_problem_at_random(difficulty)
-    if not problem:
-        return jsonify({'error': 'No problem found'}), 404
-    
-    return jsonify(problem), 200
-
-# Used by Leetcode bot to set a problem as solved
-@bp.route('/problem/<int:problem_id>/solved', methods=['PUT'])
-def mark_problem_solved(problem_id):
-    if not problem_id:
-        return jsonify({'error': 'Problem ID not provided'}), 400
-
-    # TODO: check if this function succeed or not
-    firestore_wrapper.update_last_asked_timestamp(problem_id, datetime.now())
-
-    return jsonify({'message': f'Marked {problem_id} as solved'}), 200
 
 @bp.route('/user/<int:user_id>/<difficulty>/attempted', methods=['GET'])
 def get_user_attempted(user_id, difficulty):
